Move loader examples to new module
